major upgrade of features. Now fully functional.
Added buildhmmer to allow use of hmm models in searches. Genes
are added as comma separated list, and a hmmer model is constructed
and hmmsearch used instead of phmmer to identify homologs.

RAxML only option added. To allow breaking up of analysis into
multiple steps, and to allow phylogeny construction of a novel or
edited alignment file, the RAxML only option was included. This
starts the program at the phylogeny step. Bootstrapping is not
fully integrated - it will still perform bootstraps if the -b flag
is given, but will not also do the best ML tree and incorporate the
bootstraps onto the nodes. This can be done manually, and a future
upgrade will include this feature.

exclude genes and species. You can now specify species or genes to
not include in the alignment and phylogeny. They are excluded
prior to sequence extraction, preventing unnecessary drag on
performance.

max length can be given to prevent fusion proteins from being
selected and messing up the alignment/phylogeny. Still in beta
mode, the ramifications of using this option have not been
completely thought through. Best to use only as a last resort.

graphing. Homologs are now shown un-aligned to allow visualising
of their size. The alignment is also graphed, and allows
sequences to be broken into 10 aligned regions. Each region is
colored according to the percentage of gaps in the region (but
not by the quality of the alignment - an alignment can have no
gaps but still be bad!