major upgrade to utilise dataframes and orthomcl file format
orthologs are now determined according to the orthomcl groups
output, with the script parsing the mcl output and creating
a dictionary of ortholog groups. All differentially expressed gene
files are converted to pandas dataframes, and all genes are
converted to the relevant orthomcl group. THis replaces the
old version of converting one file's gene names into the
ortholog of the other file, and allows better management of global
ortholog lists for multi-file comparisons.