added ability to calibrate orientation of experiments
because the direction of change between two conditions is
arbitrary, according to the order in which the conditions were
loaded into DESEq2 (or whatever was used to determine significance,
it may be necessary for some experiments to "reverse" the polarity
of all the expression differences. In order to facilitate this
process (without having to specify polarity for each experiment),
the 'calibrate' parameter has been added.

This parameter specifies an ortholog group that the user knows to
be consistent in its polarity among ALL experiments being compared.
By providign this group, all experiments are modified to have
the same polarity for the ortholog in this group.