Added ability to display gene names from ortholog group output
when the common ortholog groups are requested to be printed, then
it is useful to have some sort of identifier as to what kind of
ortholog each group represents. It is now possible to provide a
file that lists gene names and their names, which will be matched
to the common orthologs.

Also cleaned up and added details for some of the functions.