added sanity check and neat plotting of total concordance measures
to ensure the iterative concordance process was providing the right
results, a second quick analysis was added, that simply tallies the
number of concordant common DEGs between all species provided.

This is also graphed as a single boxplot, which is a little tidier
than the iterative reduction graph, and more useful for
presentations.