moved parameter setting from config file to CLI
since the CLI was used to specify the location of the config file
anyway, and other parameters, it made sense to maximise its utility
and assign all parameters directly to the argparser. This also led
to a more intuitive usage, without having to understand how the
config file works (or worry about correct formats, or relative vs
absolute paths).

Also clarified the help terms

Also added comments in script to clarify code.