added dynamic global analyses!
solved the major problem for the global analyses option, which was
setting the comparative statements to match the number of samples.

now all basic stats are collected (positive or negative expression,
signficant, large fold-change in expression, in all or in at
least one sample), allowing quick and easy combinatorial analysis of
these sets (such as large and positive change in all, or significant
and concordant, etc).