added heatmapping, managing duplicates and providing filter lists
it is now possible to output a heatmap of the log2(fold change)
values in the final global dataframe. Adding to this, the user can
input a list of orthologs, which will be used to filter the df,
allowing custom choice of which orthologs are used for the heatmap
(or any other global analysis).

Because pulling out orthologs from a list can involve managing
duplicates from certain species, the --manage_duplicates option
will check and selectively remove all genes from a species for a
given ortholog group, if that species hasa two or more genes in
the group (when this option is not selected, the whole ortholog is
dropped). For linkage and graphing, all null values in the df
are given a zero value.